+++ src/main/kotlin/operators/Operator.kt
@@ -1,4 +1,4 @@
-package de.rki.jfn.operator
+package de.rki.jfn.operators
 
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.node.ArrayNode
